+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Notable changes to this project are documented in this file.
 
-# Unreleased
+# 0.1.10
 ## Added
  - new config subcommand - rebuild 
  - Projector server v0.47.11
@@ -10,6 +10,7 @@ Notable changes to this project are documented in this file.
 
 # 0.1.9
 ## Added 
+ - Projector server v0.46.10
  - New command line option --cache-directory
  - Android Studio 4.2 Canary 12 compatibility
  - JVM version check